test: improve notes cleanup registration hooks
