+++ packages/core/src/module.ts
@@ -31,11 +31,9 @@ import { TimeAdvancementService } from './core/time-advancement-service';
 import type { MemoryMageAPI } from './types/external-integrations';
 import { registerSettingsPreviewHooks } from './core/settings-preview';
 import type { SeasonsStarsAPI } from './types/foundry-extensions';
-import type {
-  ErrorsAndEchoesAPI,
-  ExtendedNotesManager,
-  ExtendedCalendarManager,
-} from './types/external-integrations';
+import { registerErrorsAndEchoesIntegration } from './core/errors-and-echoes-integration';
+import { registerNotesCleanupHooks } from './core/notes-cleanup';
+import type { ExtendedNotesManager, ExtendedCalendarManager } from './types/external-integrations';
 import type {
   CalendarDate as ICalendarDate,
   DateFormatOptions,
@@ -79,145 +77,8 @@ export function setSeasonsWarningState(warned: boolean): void {
 // Register scene controls at top level (critical timing requirement)
 SeasonsStarsSceneControls.registerControls();
 
-// Register Errors and Echoes hook at top level (RECOMMENDED - eliminates timing issues)
-Hooks.once('errorsAndEchoesReady', (errorsAndEchoesAPI: ErrorsAndEchoesAPI) => {
-  // E&E is guaranteed to be ready when this hook is called
-  try {
-    Logger.debug('Registering with Errors and Echoes via hook');
-
-    errorsAndEchoesAPI.register({
-      moduleId: 'seasons-and-stars',
-
-      // Context provider - adds useful debugging information
-      contextProvider: () => {
-        const context: Record<string, unknown> = {};
-
-        // Add current calendar information - safe property access
-        if (calendarManager) {
-          const currentDate = calendarManager.getCurrentDate();
-          const activeCalendar = calendarManager.getActiveCalendar();
-
-          context.currentDate = currentDate
-            ? `${currentDate.year}-${currentDate.month}-${currentDate.day}`
-            : 'unknown';
-          context.activeCalendarId = activeCalendar?.id || 'unknown';
-          context.calendarEngineAvailable = !!calendarManager.getActiveEngine();
-        }
-
-        // Add widget state - simple property checks don't need try-catch
-        const activeWidgets: string[] = [];
-        if (CalendarWidget.getInstance?.()?.rendered) activeWidgets.push('main');
-        if (CalendarMiniWidget.getInstance?.()?.rendered) activeWidgets.push('mini');
-        if (CalendarGridWidget.getInstance?.()?.rendered) activeWidgets.push('grid');
-        context.activeWidgets = activeWidgets;
-
-        // Add system information - basic property access
-        context.gameSystem = game.system?.id || 'unknown';
-        context.foundryVersion = game.version || 'unknown';
-        context.smallTimeDetected = !!document.querySelector('#smalltime-app');
-
-        return context;
-      },
-
-      // Error filter - focus on errors relevant to S&S functionality
-      errorFilter: (error: Error) => {
-        const stack = error.stack || '';
-        const message = error.message || '';
-
-        // Always report errors that mention our module explicitly
-        if (
-          stack.includes('seasons-and-stars') ||
-          message.includes('seasons-and-stars') ||
-          message.includes('S&S') ||
-          stack.includes('CalendarManager') ||
-          stack.includes('CalendarWidget') ||
-          stack.includes('CalendarEngine') ||
-          stack.includes('NotesManager')
-        ) {
-          return false; // Don't filter (report this error)
-        }
-
-        // Report time/calendar related errors that might affect us
-        if (
-          message.includes('worldTime') ||
-          message.includes('game.time') ||
-          message.includes('calendar') ||
-          message.includes('dateToWorldTime') ||
-          message.includes('worldTimeToDate') ||
-          (message.includes('time') && stack.includes('foundry'))
-        ) {
-          return false; // Don't filter (time system errors affect us)
-        }
-
-        // Report widget positioning and UI errors
-        if (
-          message.includes('widget') ||
-          message.includes('SmallTime') ||
-          message.includes('player list') ||
-          (message.includes('position') && stack.includes('ui')) ||
-          message.includes('ApplicationV2')
-        ) {
-          return false; // Don't filter (UI errors might affect our widgets)
-        }
-
-        // Report integration-related errors
-        if (
-          message.includes('Simple Calendar') ||
-          message.includes('simple-calendar') ||
-          message.includes('compatibility') ||
-          message.includes('bridge') ||
-          stack.includes('integration')
-        ) {
-          return false; // Don't filter (integration errors affect us)
-        }
-
-        // Report foundry core time system errors
-        if (
-          stack.includes('foundry.js') &&
-          (message.includes('time') || message.includes('world') || message.includes('scene'))
-        ) {
-          return false; // Don't filter (core time system issues)
-        }
-
-        // Filter out errors from unrelated modules (unless they mention calendar/time)
-        const unrelatedModules = [
-          'dice-so-nice',
-          'lib-wrapper',
-          'socketlib',
-          'combat-utility-belt',
-          'enhanced-terrain-layer',
-          'token-action-hud',
-          'foundryvtt-forien-quest-log',
-        ];
-
-        for (const module of unrelatedModules) {
-          if (
-            stack.includes(module) &&
-            !message.includes('calendar') &&
-            !message.includes('time') &&
-            !stack.includes('seasons-and-stars')
-          ) {
-            return true; // Filter out (unrelated module error)
-          }
-        }
-
-        // Default: filter out most other errors unless they seem time/calendar related
-        if (message.includes('calendar') || message.includes('time') || message.includes('date')) {
-          return false; // Don't filter (might be related)
-        }
-
-        return true; // Filter out everything else
-      },
-    });
-
-    Logger.debug('Successfully registered with Errors and Echoes via hook');
-  } catch (error) {
-    Logger.error(
-      'Failed to register with Errors and Echoes via hook',
-      error instanceof Error ? error : new Error(String(error))
-    );
-  }
-});
+// Register Errors and Echoes integration separately
+registerErrorsAndEchoesIntegration(() => calendarManager);
 
 /**
  * Module initialization
@@ -288,7 +149,7 @@ Hooks.once('ready', async () => {
   await notesManager.initialize();
 
   // Register notes cleanup hooks for external journal deletion
-  registerNotesCleanupHooks();
+  registerNotesCleanupHooks(notesManager);
 
   // Register with Memory Mage if available
   registerMemoryMageIntegration();
@@ -1710,64 +1571,6 @@ function getActiveWidgetCount(): number {
   if (CalendarGridWidget.getInstance?.()?.rendered) count++;
 
   return count;
-}
-
-/**
- * Register hooks to clean up notes when journals are deleted externally
- */
-function registerNotesCleanupHooks(): void {
-  // Hook into journal deletion to clean up our notes storage
-  Hooks.on(
-    'deleteJournalEntry',
-    async (journal: JournalEntry, _options: Record<string, unknown>, _userId: string) => {
-      Logger.debug('Journal deletion detected', {
-        journalId: journal.id,
-        journalName: journal.name,
-        isCalendarNote: !!journal.flags?.['seasons-and-stars']?.calendarNote,
-      });
-
-      try {
-        // Check if this was a calendar note
-        const flags = journal.flags?.['seasons-and-stars'];
-        if (flags?.calendarNote) {
-          Logger.info('Calendar note deleted externally, cleaning up storage', {
-            noteId: journal.id,
-            noteName: journal.name,
-          });
-
-          // Remove from our storage system
-          if (notesManager?.storage) {
-            await notesManager.storage.removeNote(journal.id);
-            Logger.debug('Note removed from storage');
-          }
-
-          // Emit our own deletion hook for UI updates
-          Hooks.callAll('seasons-stars:noteDeleted', journal.id);
-
-          // Refresh calendar widgets to remove the note from display
-          const calendarWidget = CalendarWidget.getInstance?.();
-          if (calendarWidget?.rendered) {
-            calendarWidget.render();
-          }
-          const miniWidget = CalendarMiniWidget.getInstance?.();
-          if (miniWidget?.rendered) {
-            miniWidget.render();
-          }
-          const gridWidget = CalendarGridWidget.getInstance?.();
-          if (gridWidget?.rendered) {
-            gridWidget.render();
-          }
-        }
-      } catch (error) {
-        Logger.error(
-          'Failed to clean up deleted calendar note',
-          error instanceof Error ? error : new Error(String(error))
-        );
-      }
-    }
-  );
-
-  Logger.debug('Notes cleanup hooks registered');
 }
 
 // Export functions for testing
