+++ packages/core/src/ui/calendar-grid-widget.ts
@@ -376,7 +376,7 @@ export class CalendarGridWidget extends foundry.applications.api.HandlebarsAppli
         }
       } catch (error) {
         // Silently handle moon calculation errors to avoid breaking calendar display
-        console.debug('Error calculating moon phases for date:', dayDate, error);
+        Logger.debug('Error calculating moon phases for date:', { dayDate, error });
       }
 
       currentWeek.push({
